.history/requirements_20260122100344.txt
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==3.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyMySQL==1.0.2
SQLAlchemy==1.4.29
Werkzeug==2.0.2
requests==2.28.1
cryptography==3.4.8
python-dotenv==0.19.2
wechatpy==1.8.14
pycryptodome==3.20.0
